scripts/requirements.txt
colorama==0.4.6
pyinstaller==6.5.0
requests==2.31.0
psutil==5.9.8
PyJWT==2.8.0
python-dotenv==1.0.1
